my-app/requirements.txt
discord.py
python-decouple
tensorflow
numpy
requests